// src/app/modules/matenimiento/models/producto/productoResponseDetalle.model.ts
export class ResponseDetalleProcedureProducto {
    idModelo: number = 0;
    idProducto: number = 0;
    color: string ="";
    categoria: string ="";
    talla: string ="";
    material: string ="";
    descripcion: string ="";
    fotografia2: string ="";
    fotografia3: string ="";
    fotografia4: string ="";
    idDetalleProducto: number = 0;
    nombreProd: string ="";
    precioUnitario: number =0;
}